Use int instread of byte for max priority. Seems like Rabbit has changed from byte to int.

+++ src/NServiceBus.Transport.RabbitMQ/Configuration/RabbitMQTransportSettingsExtensions.cs
@@ -34,7 +34,7 @@
         /// </summary>
         /// <param name="transportExtensions"></param>
         /// <param name="maxPriority">Max priority on queue. 0 - 9</param>
-        public static TransportExtensions<RabbitMQTransport> UseConventionalRoutingTopologyWithPriority(this TransportExtensions<RabbitMQTransport> transportExtensions, byte maxPriority) =>
+        public static TransportExtensions<RabbitMQTransport> UseConventionalRoutingTopologyWithPriority(this TransportExtensions<RabbitMQTransport> transportExtensions, int maxPriority) =>
             transportExtensions.UseRoutingTopology(durable => new ConventionalRoutingTopology(durable, maxPriority));
